requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.9
numpy>=1.24.0
fastapi==0.129.0
uvicorn==0.41.0
python-multipart==0.0.22
PyAudio==0.2.14
ollama
python-dotenv
httpx
spacy
nltk
xgboost
scikit-learn
torch
networkx

transformers==4.41.2
tokenizers==0.19.1
huggingface-hub==0.24.7
safetensors==0.4.4
faiss-cpu==1.8.0
datasketch==1.6.5
pdfplumber==0.11.4
pdf2image==1.17.0
easyocr
Pillow==10.4.0
accelerate==0.33.0
google-genai
rake_nltk
moviepy
openai-whisper
groq
requests
yt-dlp

# Django + DRF
django>=4.2
djangorestframework>=3.14

# Redis caching (optional — falls back to in-memory if unavailable)
redis>=5.0
